wip: rename accordion-section component files
