+++ src/app/shared/components/template/components/layout/accordion-section/accordion-section.component.html
@@ -1,0 +1,48 @@
+<div class="accordion-wrapper">
+  <div
+    [ngClass]="{
+      completed: completed && !_row.disabled,
+      disabled: _row.disabled,
+      inProgress: !completed && !_row.disabled && percentComplete > 0,
+      notStarted: !completed && !_row.disabled && percentComplete === 0
+    }"
+    class="accordion-status"
+  >
+    <img
+      *ngIf="completed && !_row.disabled"
+      class="tick-icon"
+      [src]="'plh_images/icons/tick.svg' | plhAsset"
+    />
+    <img *ngIf="_row.disabled" [src]="'plh_images/icons/temporarily_disabled.svg' | plhAsset" />
+    <img
+      *ngIf="!completed && !_row.disabled"
+      [src]="'plh_images/icons/in_progress.svg' | plhAsset"
+    />
+  </div>
+  <div
+    class="accordion-section"
+    [ngClass]="{
+      openSection: _row.parameter_list.state === 'open',
+      disabled: _row.disabled,
+      completed: completed,
+      inProgress: !completed && !_row.disabled
+    }"
+  >
+    <div class="progress" [ngStyle]="{ width: percentComplete + '%' }"></div>
+    <h2 (click)="toggleOpen()">{{ title }}</h2>
+    <div
+      class="accordion-section-content"
+      [ngClass]="{
+        openSection: _row.parameter_list.state === 'open'
+      }"
+    >
+      <plh-template-component
+        style="z-index: 2"
+        *ngFor="let childRow of _row.rows; trackBy: trackByRow"
+        [row]="childRow"
+        [parent]="parent"
+      >
+      </plh-template-component>
+    </div>
+  </div>
+</div>
