+++ src/app/shared/components/template/components/layout/workshops_accordion.ts
@@ -16,7 +16,6 @@ import { TemplateBaseComponent } from "../base";
       </plh-tmpl-accordion-section>
     </div>
   </div>`,
-  styleUrls: ["./accordion.scss"],
 })
 export class WorkshopsComponent extends TemplateBaseComponent implements OnInit {
   rows: FlowTypes.TemplateRow[];
